Catalogs in java now maintain subversion-esque version numbers for all nodes. These numbers are incremented on all changed nodes for every call to catalog.execute(...). They apply only for the current instance of a catalog, and any deep-copied copies of the catalog. If you serialize and de-serialize, then all versioning info is reset.
This allows you to apply a catalog diff to a catalog, then inventory what has changed by examining the data structure. Initially, this will be used to limit changes to auth + procs, but eventually it could be used to make smaller changes to the execution environment for small catalog changes. For now, all procs and auth are reloaded for every change in every EE. This works ok for procs, but is pretty lousy once we support changing tables.