ENG-149: Dumps all plan nodes in EE if SIGSEGV is caught.
This uses Sun's Java signal chaining, so it only works with Sun's JVM. In order
to use it, you must set LD_PRELOAD to the libjsig.so library that comes with
your JRE.

On systems other than linux, none of this code will be compiled in. If it is run
without LD_PRELOAD set or in a different JVM other than Sun's, the function is a
no-op.