Fix ENG-450:
Modify the PlanColumns generated for AggregateNode output schema to
contain an appropriately-configured TupleValueExpression rather than
the AggregateExpression.

Modify calculateTupleValueColumnIndexes so that it special-cases
display columns that are aggregate expressions in order to look them
up properly and replace the AggregateExpression in the display column
with the matching TupleValueExpression from the aggregate node output
for that aggregated column.

Add some test cases for the planner and order by regression suites.

Add order by to the results stored procedure in the examples/voter app.

(Oh, the hackery)