ELClient refactoring checkpoint:
Refactor the EL client code to work with multiple nodes.
- Enable VoltDB app and the regression suite code to support ELT tests
  with multiple nodes using LocalCluster.

This is at least as functional as the version currently committed and
much much closer to its final form.  Should work with all cluster
configs with the exception of replication which is currently still
broken.  Committing this checkpoint to unblock rbetts.