ENG-596
Trim the columns extracted from persistent tables to only those that are
used somewhere in the plan.

This patch is a major rewrite to column tracking in the planner and
the EE.  Every plan node that used to use an explicit column index
into a tuple and then discover that index during initialization in the
EE has been updated to use an abstract expression which has the table
names and column indexes updated by the planner.