Connect src and dest tables when adding table with new materialized view.
Noticed this missing testcase when reviewing my add/drop table diff
and sure enough the corresponding functionality wasn't quite
right. When adding a table with a view, the view table was created and
all SQL would run but the source table would not push tuples to the
view as the source and destination tables weren't fully connected.