BenchmarkClient now calls compileAllCatalogs() on client project builder.
Before, benchmark reflected static string to discover the desired
catalog name. Now it invokes a new VoltProjectBuilder method:
compileAllCatalogs(). This allows the project builder to compile
multiple catalogs. It returns a list of strings that are catalog
filenames.  Benchmark client will scp all of these to the remote hosts
and use the first catalog in the list to start the benchmark run.

This should allow a benchmark to compile multiple catalogs to use via
@UpdateApplicationCatalog.