Moves the snapshot file name constructor to SnapshotUtil so that it can be reused.
Move snapshot status under the management of StatsAgent.

Include the snapshot file name in the result of snapshotstatus. It is hard for
the client to figure out the name of a snapshot file without asking the server
to do a snapshot scan or constructing a Table catalog object. Including this bit
of information makes it much easier for the client to get the name of a snapshot
file.

Turning the SnapshotStatus sysproc to use the StatsAgent to get results.

Change the TestSaveRestoreSysprocSuite test to use the new SnapshotStatus
result.

Dedicate a method to generate the snapshot digest file name and make sure that
all places use it.

Isolate the CRC check into a static method so that it can be reused.