ENG-606
This is not a fix for ENG-606 but a patch to gather some additional
data from Hudson if this issue occurs again.

Increase out-of-process planner unresponsive timeout to a minute.  If
this occurs, it had better be for a reason other than a loaded hudson
machine.

Some minor tweaks to the OPP logging.  The volt server will now report
whether or not the OPP is still running when it becomes unresponsive,
what the exit code was if it died, and what the last SQL it attempted
to plan was if it hung.  This data sort-of goes to plannerlog.txt, but
it's inconvenient to try to find for a random failed Hudson run.