ENG-720: NullPointerException when trying to start server with no users
Fixes the NPE and adds a test case to check for it in the future.

I also changed the part of VoltProjectBuilder that creates the deployment file to -not- add the <users /> element if there are no users. This was necessary for the new test case to work and it seems like it would have been a bad thing for any test cases to be relying on this anyway.