Make licensecheck directly usable as commit hook.
Add a 'basepath' variable and pre-pend this to all paths so
the tool no longer assumes paths relative to eng/tools.

Add a ascommithook flag that can be set to true to disable
out-of-tree checks, which aren't relevant to a commit hook.

Add a comment at the top of the file with instructions for
use as commit hook.