Fix for ENG-737: "Rejoining a node with wrong deployment file succeeds erroneously"
We now can calculate the checksum of ***the contents*** of a deployment XML file, rather than the actual bytes of said XML file. We can then keep track of this checksum and verify it is the same in socketjoiner. You can't rejoin if you have a semantically different deployment file anymore.