Fixing ENG-769: "Handle the time-moving-backwards initiator error with more forgiveness"
It will now log a lot more when this badness happens (both to log4j and partially to stderr).
But if the negative time adjustment is less than 3secs, it will just wait.
Note, this threshold is a constant in VoltDB. It could be user configurable if we decide to do that later.