Adding a test that is going to be awesome for ENG-712: "Rejoin must synchronize export stream metadata", but isn't done. But the test currently is nice for ENG-782: "Single node recovery unit tests lack parallelism to necessary to reveal issues."
Unlike the rejoin fuzz, it creates a slew of clients that hammer a cluster that can't seem to keep all of its nodes online. I had to actually think about how to do this, as our clients need some work to make them easy to reconnect to failed nodes when they return.

The tests runs and does as much damage as it can in 60 seconds. It can be run forever, but 60 seconds is the default.