ENG-798 Correctly map view::source columns
Views that re-ordered their source table columns
were not correctly configured. The order of the
re-mapped source columns was not maintained.

That is, given the source table C1, C2 and
the view group-by schema: C2, C1, the view configuration
mapped source column C1 to the first column of the
view's schema and column C2 to the second column of the
view's row schema.