ENG-879
The bulk loading used for restore doesn't properly support undo, which
was accounted for on the base table.  However, when you add views, by
the time we update the view, we don't know we're doing a bulk load and
we create lots of undo quantums.  We weren't properly releasing the
undo token we were creating for the bulk load operation, and so we
were accumulating a bunch of memory with every call.  This patch
actually releases those tokens.