This patch removes the contiguous fast allocator that was used for allocating the free list in tuple block.
FastAllocator is modified so that requests for contiguous allocations go
to the heap. I think this makes more sense then using a linear time
object pool for contiguous allocations (not a strong point). Tree
indexes and strings aren't contiguous allocations so they pool nicely.
Hash indexes do contiguous allocations for the buckets and the deque
used for the free list also does a contiguous allocation for the map of
blocks.

Change the free list to in TupleBlock.h to use a deque, I checked the
implementation and it doesn't require contiguous memory and does some
allocation that is pool friendly and some allocation that is contiguous
and not pool friendly.

Create a custom allocator for use by the the boost pools served by
thread local pool that maintains a thread local count of memory
allocated for pools and add a statistic accessible via node memory for
it. The thread local value is retrieved along with the other node memory
statistics via a JNI method call.

Move fast allocator to third party under a new boost extensions
directory. Fast allocator is based on code from boost pool_alloc. Put
the boost license in the header file along with our X11 copyright.

r=rbetts