This commit contains a range of changes.
There are changes to the project.xml and deployment.xml that move configuration of exports enabled/disabled status from project to deployment. This relates to ENG-972 and ENG-971.

This also fixes ENG-969, the counter is back and accurately depicts how many bytes are left to be pulled from each export stream. It is a little rough once overflow to disk starts because it releases a file at a time and the files are 64 megs and must be completely emptied before being deleted.

Export overflow can now function as a persistence mechanism. If quiesce is invoked then all export buffers will be written to disk and synced. If snapshot save is invoked all the export buffers will be written disk, but not synced. On startup a Volt process looks for export data that has been persisted to disk and serves it up to export clients. For 1.3 export data should not be moved between servers. Adding servers is fine, but data from removed servers can't be moved to a remaining server's export overflow directory.

For testing there is now an overloaded version of java.io.File called VoltFile that adds a prefix at most once to all paths passed to it. The at most once property is maintained by adding a bit of magic to the prefix and checking for its presence before prepending the prefix.

This prefix can be set in process or at the time a process is started via a property. This is used by local cluster to give each process in the local cluster a dedicated subroot in /tmp/#{username}. Local cluster has methods for deleting and list files/directories across subroots. VoltFile has helper methods for creating new subroots, deleting all subroots etc. This makes it possible to control the visibility of files when running multiple process via local cluster.

Don't do a synchronous flush of export data when initiating a snapshot.