ENG-1047:
Fix an issue that would lead to an inconsistent table iterator in
RecoveryContext.  We used to initialize the iterator in the
constructor, but transactions could sneak in between the constructor
and the calls to nextMessage() which would cause the table state to
change which would not be reflected in the iterator.