ENG-732: Push-down count(*).
count(*) will now be pushed down in the following cases,

- count(*) is the only select column.
- count(*) with other group by columns.

For the above cases, the explain plan will include a count(*) aggregate right
above the scan, followed by a send to the coordinator, followed by a sum
aggregate.

Make PlannerTestAideDeCamp's compile() method return all the fragments, not just
the first fragment, so that for distributed plans we can look at all the stuff.

Fixed a trivial Java doc error.