Fixing ENG-1198: "Java client auth code doesn't support partial socket reads everywhere"
We now support partial reads in more places. There isn't an explicit test for this, but it's pretty well covered by other code. The test I'm going to check in for ENG-1060 can hit it occasionally, but it's difficult to reproduce manually (i.e. we haven't ever hit it until this week in all of our testing).

I think a code review makes more sense than someone trying to reproduce this.