Test robustness fix:
John's machine believes that there are occasionally only 2,999,998,000
nanoseconds in 3 seconds.  There's no reason why these assertions have
to be so tight to 3 seconds; we just care that they're significantly
above MAX_INT to check for overflow.  

Also, since we're here, give a meaningful error message if the
assertion fails rather than the ever-popular-and-oh-so-useful
"junit.framework.AssertionFailedError: null"