PRO-421
Add an enabled flag to SnapshotSchedule.  Always create an internal
SnapshotSchedule, populate it with the snapshot path, and use the
enabled flag (rather than presence of this element) in the
SnapshotSchedule in a running catalog to determine whether or not auto
snapshots are enabled.  This behavior is consistent with command
logging and export.