Fix SQLCommand.
Prepopulate the parameter array to make sure that the array size is equal to
parameter count. The procedure columns returned by @SystemCatalog are not sorted
by their ordinal positions. So adding them into the array requires the array
size to be correct in the first place, or an IndexOutOfBoundException will be
thrown.