ENG-1832: Add invocation acceptance policy to secondary cluster.
Add basic invocation acceptance policies to the ClientInterface. When it
receives a procedure invocation, it will check the invocation against a set of
policies to determine if it should accept it. If any of the policy says no, the
invocation is rejected.

For now, there is only one policy on secondary cluster to reject writes from
normal clients and reads from DR agent.

A lot of the checkings in handleRead() may be converted to policies later.