Critical fix for command logging. SnapshotDaemon sometimes will delete the wrong snapshot.
TreeMap.Entry for an iterator of a descendingMap.entryset is invalided by iterator.remove.
If only two items are present TreeMap.Entry points to the correct entry. Otherwise it points to a different entry
potentially resulting in the wrong truncation snapshot being deleted by the snapshot daemon.

This doesn't cause data loss if it occurs once, but it does prevent replay without manual intervention. If it were to occur multiples times it could cause data loss.

This fix moves the iterator.remove() to after entry.getValue() is invoked and adds logging to make the actions of the snapshot daemon less opaque.