Fix several issues, Periodic Work thread was doing all tasks on shutdown, SnapshotDaemon had an extra 10 second delay that isn't necessary, bug in agreement site.
The periodic work thread was a ScheduledExecutorService and on shutdown it was executing all periodic and delayed tasks. Created a factory method in MiscUtils to create the executor service and automatically disable executing tasks on shutdown.

SnapshotDaemon had the ten second gathering period in two places, the second wasn't necesary and was diruptive in TestCommandLogServer.

Fix a bug causing spurious failures in TestCommandLogServer. AgreementSite on rejoin was not filtering out txns from before the txnid of the snapshot that is uses during rejoin.