For ENG-3243, in TestSnapshotDaemon check async condition in a spin loop instead of a fixed timeout
I think we had a bad instance for a while and that caused the fixed timeout to be insufficient.

I converted it to a spin loop with a large timeout.

I was unable to reproduce the issue with an m2.xlarge that I spun up. This is both before and after the fix which is the reason I suspect a bad instance.