Rethread async compiler to process planned transactions and catalog update in the network thread of the requesting client.
This solves two issues. In IV2 there is network thread local state that needs to be accessed lock free from the network thread.

In pre-IV2 Volt this change is beneficial because you no longer rely on the periodic work thread to come every 5 milliseconds and pick up the finished compiler work.