ENG-3374: Send the list of non-coordinator sites from the initiator to the multi-part coordinator.
During rejoin, the initiator's site tracker and the coordinator's site tracker
may be inconsistent for a brief period of time. The initiator needs to pass the
set of non-coordinator sites it has sent participant notices to to the
coordinator, so that when the coordinator sends out fragment works, all
participating sites will get them. Otherwise, the cluster will deadlock on
rejoin.