Fix a bug in RestoreAgent where it read an invalid JSON document
This shows up in TestRestoreAgent. It is a ZK node with a document that is created under a directory while the directory is being parsed by restore agents that are only expecting to find restore information documents. Presumably the leader is getting ahead of the other restore agents and creating the snapshot_id document before they are done scanning.

The fix is to skip the one node that is created by the leader called snapshot_id in RestoreAgent.deserializeRestoreInformation