For ENG-2543, implement short circuit reads where replicas execute read only single parts
This includes server mods to allow the client interface to route to local replicas and a mode to replicas
to allow them to run single part read transactions when they are not the leader.

The client library uses the metadata from the server to determine which procedures are reads and will
load balance them across replicas.