ENG-4104: Provide proper barrier to snapshot initialization.
To avoid having the SnapshotCompletionMonitor race with the snapshot
participants during snapshot initialization, the MPI which runs the SnapshotSave
sysproc gets the participating host count based on the hosts in the "hosts"
field in the snapshot completion ZK node. When each node finishes taking
snapshot, it decrements the host count. The SnapshotCompletionMonitor monitors
the host count and fires completion interests when this number reaches 0.